alternating maximization

+++ rsnn/firing_sequences/utils.py
@@ -27,13 +27,15 @@ def is_predictable(firing_sequences, Tr):
     if B > 1:
         raise NotImplementedError
 
+    firing_sequences = firing_sequences.double()
+
     # cyclic condition
-    padding1 = F.pad(firing_sequences, (0, Tr - 1), mode="circular").double()
-    padding2 = F.pad(firing_sequences, (0, Tr), mode="circular").double()
+    padding1 = F.pad(firing_sequences, (0, Tr - 1), mode="circular")
+    padding2 = F.pad(firing_sequences, (0, Tr), mode="circular")
 
     # one-to-one correspondence from L-channels windows of length T to [0, (T+1)^L - 1], where T = Tr or Tr + 1.
-    filter1 = torch.arange(1, Tr + 1, dtype=float)[None, None, :] * torch.pow(Tr + 1, torch.arange(L, dtype=float))[None, :, None]
-    filter2 = torch.arange(1, Tr + 2, dtype=float)[None, None, :] * torch.pow(Tr + 2, torch.arange(L, dtype=float))[None, :, None]
+    filter1 = torch.arange(1, Tr + 1, dtype=torch.double)[None, None, :] * torch.pow(Tr + 1, torch.arange(L, dtype=torch.double))[None, :, None]
+    filter2 = torch.arange(1, Tr + 2, dtype=torch.double)[None, None, :] * torch.pow(Tr + 2, torch.arange(L, dtype=torch.double))[None, :, None]
 
     convolution1 = F.conv1d(padding1, filter1, groups=1)
     convolution2 = F.conv1d(padding2, filter2, groups=1)
@@ -51,13 +53,15 @@ def is_predictable(firing_sequences, Tr):
 def count_predictable(firing_sequences, Tr):
     B, L, _ = firing_sequences.size()
 
+    firing_sequences = firing_sequences.double()
+
     # cyclic condition
-    padding1 = F.pad(firing_sequences, (0, Tr - 1), mode="circular").double()
-    padding2 = F.pad(firing_sequences, (0, Tr), mode="circular").double()
+    padding1 = F.pad(firing_sequences, (0, Tr - 1), mode="circular")
+    padding2 = F.pad(firing_sequences, (0, Tr), mode="circular")
 
     # one-to-one correspondence from L-channels windows of length T to [0, (T+1)^L - 1], where T = Tr or Tr + 1.
-    filter1 = torch.arange(1, Tr + 1, dtype=float)[None, None, :] * torch.pow(Tr + 1, torch.arange(L, dtype=float))[None, :, None]
-    filter2 = torch.arange(1, Tr + 2, dtype=float)[None, None, :] * torch.pow(Tr + 2, torch.arange(L, dtype=float))[None, :, None]
+    filter1 = torch.arange(1, Tr + 1, dtype=torch.double)[None, None, :] * torch.pow(Tr + 1, torch.arange(L, dtype=torch.double))[None, :, None]
+    filter2 = torch.arange(1, Tr + 2, dtype=torch.double)[None, None, :] * torch.pow(Tr + 2, torch.arange(L, dtype=torch.double))[None, :, None]
 
     convolution1 = F.conv1d(padding1, filter1, groups=1)
     convolution2 = F.conv1d(padding2, filter2, groups=1)
